Changes the behavior of query string construction. Previously I used a white-list of allowed query string variable names. Any of the allowed names that were defined in config were incorporated into the final query string. This is problematic because there's no such thing as a complete list of variable names for query strings.
So the solution is to make query string variables work the way that POST variables work. Config
vars that need to go into the query string of the request are defined as "$this->config['qs']['var']='value'.

Cleaned up the code related to this change and updated the example job files to make use of it.