Bugzilla bug #15
Fix a problem with loading custom config files and the error messages show when
config files are not found.

The problem was that if the user specified a custom config file, say 'my.custom.config.php',
and that file is only a partial, and there's no correspondingly named config files in
the main config directory, the Config class won't load the default config file as a base.

The main change here is to always load the default config file, to form a base of configurations
for custom files to selectively overwrite.

The other changes are to improve error detection and error messages. It's OK to specify no
config file, it's ok to specify a partial config file, it's NOT ok to specify a non-existent
config file.