Refactor the Expectations Engine.
Set up a base abstract class and an interface for Expectation Testers.
Move all of the tests out of the Expectations Engine class and into their own classes.
Change the Expectation Engine to load tester classes and run their test() method
for every expectation specified in a given job.
Create some new Exceptions for missing expectation files and classes.
Create an exmaple of using expectations in a job.