Formatter Refactor and clean-up.
Moved the formatter classes into their own directory.
Changed the formatter classes namespaces to reflect that new directory.
Updated the harness's formatter factory method to load the formatters properly.
Updated the harness's autoload method to leverage the "namespace matches directory"
pattern so that the autoloader looks in a lib sub directory based on the namespace
of the class it wants to load.
Updated the Randomizer and Expectation classes so that their namespace matches
their directory location.
Fixed a few namespace references in the application to match the convention of
namespace matches directory.
Removed many calls to require_once() that were no longer needed due to updates
in the harness's autoload method.