Added some code for setting expectations on a job programmatically. This allows JobSeries to set expectations on Jobs as it works through the series.
Also put in some special handling around setting the expected HTTP return code value.

If an expectation sets the expected value for the http return code, receiving that code
should not trigger an exception. Receiving a different code should. However, a 200 should
never trigger an exception because when it does, we don't see the results of the request.
Those could be very informative and should be displayed.