Made the Reconnector's base property configurable instead of hard coded, so the harness can pretend to be a mobile client.
Added a clearToken() method to the RestConnector

Updated the JobAbstract to accept http return codes lower than 300 instead of only equal to 200.

Added timestamps to print the start, finish and elapsed times for a given harness run.

updated the rest version in the config file.

updated Upload and Install demo jobs